backend/requirements.txt
# Backend dependencies for AITradingAdvisory Web API

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# WebSocket
websockets>=12.0

# Security & Auth
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # Pin to avoid passlib compatibility issues

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Async
aiofiles>=23.2.0
httpx>=0.26.0

# Include existing dependencies from main project
autogen-agentchat>=0.4.0
autogen-ext[magentic-one]>=0.4.0
openai>=1.12.0
tiktoken>=0.5.0
requests>=2.31.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
rich>=13.7.0
pyyaml>=6.0.1
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
kaleido>=0.2.1
